use visitor pattern to support different output formats
Instead of hard-coding the output format, use callbacks to notify a handler of the different relevant parts of the page. Currently only affects the CLI version, arguments and output remain unchanged. The resulting code is tested to be python 2 and 3 compatible as long as the user set an appropriate PYTHONIOENCODING for German command line output from python2.