use visitor pattern to support HTML output
Removed the duplicated HTML-traversal code.
Implemented HTML formatting as handler and reused these in the CGI-script.
The CLI script can now also generate HTML output by using the "--html switch".
Both scripts are tested to be compatible with both python 2 and 3, the output of the CGI is always UTF-8, no matter the locale set by the environment/request.